another minor adjustment:
if we log out, we make a note of the time that happens, and keep track of it
for 600 seconds.

if that logout time is greater than the modification time of the document,
we prefer it when calculating whether to send not_modified.

--HG--
extra : convert_revision : svn%3A2b5a5965-8a46-0410-aa19-a76fe43533fc/trunk%402414